requirements.txt
# Core
flask==3.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-socketio==5.3.5
python-socketio==5.10.0
eventlet==0.33.3

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Web Scraping
selenium==4.15.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.2

# Data Processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
requests==2.31.0
schedule==1.2.0
psutil==5.9.6
waitress==2.1.2

# Development
ipython==8.18.1
rich==13.7.0

# Optional
# redis==5.0.1
# celery==5.3.4
# gunicorn==21.2.0
